iota_stardust_types

Module block

Core data types for blocks in the tangle.

Re-exports

pub use self::payload::MilestoneOption;

pub use self::payload::ParametersMilestoneOption;

pub use self::protocol::PROTOCOL_VERSION;

pub use self::protocol::ProtocolParameters;

Modules

address

A module that provides types and syntactic validations of addresses.

output

A module that provides types and syntactic validations of outputs.

payload

Payload types. Payload types.

protocol

Protocol parameters. Protocol parameters.

Structs

BlockId

A block identifier, the BLAKE2b-256 hash of the block bytes. See https://www.blake2.net/ for more information.

MilestoneId

A milestone identifier, the BLAKE2b-256 hash of the milestone bytes. See https://www.blake2.net/ for more information.

MilestoneIndex

A wrapper around a u32 that represents a milestone index.

TransactionId

A transaction identifier, the BLAKE2b-256 hash of the transaction bytes. See https://www.blake2.net/ for more information.

Enums

Error

Error occurring when creating/parsing/validating blocks.

Traits

ConvertTo