iota_stardust_types::block

Module output

A module that provides types and syntactic validations of outputs.

Re-exports

pub use self::alias::AliasOutput;

pub use self::alias::AliasOutputBuilder;

pub use self::alias::AliasTransition;

pub use self::basic::BasicOutput;

pub use self::basic::BasicOutputBuilder;

pub use self::feature::Feature;

pub use self::feature::Features;

pub use self::foundry::FoundryOutput;

pub use self::foundry::FoundryOutputBuilder;

pub use self::nft::NftOutput;

pub use self::nft::NftOutputBuilder;

pub use self::unlock_condition::UnlockCondition;

pub use self::unlock_condition::UnlockConditions;

Modules

alias

basic

feature

foundry

nft

unlock_condition

Structs

AliasId

Unique identifier of an alias, which is the BLAKE2b-256 hash of the Output ID that created it.

FoundryId

Defines the unique identifier of a foundry.

NativeToken

NativeTokens

NativeTokensBuilder

A builder for NativeTokens.

NftId

Unique identifier of an NFT, which is the BLAKE2b-256 hash of the Output ID that created it.

OutputId

The identifier of an Output.

SimpleTokenScheme

TokenId

Unique identifiers of native tokens. The TokenId of native tokens minted by a specific foundry is the same as the FoundryId.

TreasuryOutput

TreasuryOutput is an output which holds the treasury of a network.

Enums

ChainId

Output

A generic output that can represent different types defining the deposit of funds.

TokenScheme

Constants

OUTPUT_COUNT_MAX

The maximum number of outputs of a transaction.

OUTPUT_COUNT_RANGE

The range of valid numbers of outputs of a transaction .

OUTPUT_INDEX_MAX

The maximum index of outputs of a transaction.

OUTPUT_INDEX_RANGE

The range of valid indices of outputs of a transaction .