iota_stardust_types::block::output

Module feature

Structs

Attribute

FeatureFlags

A bitflags-based representation of the set of active Features.

Features

Irc27Metadata

The IRC27 NFT standard schema.

Irc30Metadata

The IRC30 native token metadata standard schema.

IssuerFeature

Identifies the validated issuer of the UTXO state machine.

MetadataFeature

Defines metadata, arbitrary binary data, that will be stored in the output.

SenderFeature

Identifies the validated sender of an output.

TagFeature

Makes it possible to tag outputs with an index, so they can be retrieved through an indexer API.

Enums

Feature