iota_stardust_types::block::output

Module unlock_condition

Structs

AddressUnlockCondition

Defines the Address that owns this output, that is, it can unlock it with the proper Unlock in a transaction.

ExpirationUnlockCondition

Defines a unix time until which only Address, defined in Address Unlock Condition, is allowed to unlock the output. After or at the unix time, only Return Address can unlock it.

GovernorAddressUnlockCondition

Defines the Governor Address that owns this output, that is, it can unlock it with the proper Unlock in a transaction that governance transitions the alias output.

ImmutableAliasAddressUnlockCondition

Defines the permanent AliasAddress that owns this output.

StateControllerAddressUnlockCondition

Defines the State Controller Address that owns this output, that is, it can unlock it with the proper Unlock in a transaction that state transitions the alias output.

StorageDepositReturnUnlockCondition

Defines the amount of IOTAs used as storage deposit that have to be returned to the return Address.

TimelockUnlockCondition

Defines a unix timestamp until which the output can not be unlocked.

UnlockConditionFlags

A bitflags-based representation of the set of active UnlockConditions.

UnlockConditions

Enums

UnlockCondition